requirements.txt
altgraph==0.17.2
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.2
numpy-financial==1.0.0
openpyxl==3.0.9
pandas==1.4.0
pycodestyle==2.8.0
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
